[agent_builder/capabilities/knowledge] - refactor: move data source components to knowledge capability
 - Renamed and moved various data source components to reside under the agent_builder capabilities for knowledge
 - Updated imports in `KnowledgeConfigurationSheet.tsx` to reflect the new location of `DataSourceBuilderSelector` component
