+++ front/components/agent_builder/capabilities/knowledge/DataSourceCategoryBrowser.tsx
