+++ front/components/agent_builder/capabilities/knowledge/DataSourceNodeTable.tsx
