+++ front/components/agent_builder/capabilities/knowledge/DataSourceViewTable.tsx
